WANIPConnection 0.85 changes. More argument value checking
git-svn-id: http://us000253/svn/IGDv2@370 55a18753-f1c3-4722-b33a-e097f9cf6a53

+++ linuxigd2/src/gatedevice.h
@@ -106,7 +106,7 @@ int ExpirationTimerThreadShutdown(void);
 int ScheduleMappingExpiration(struct portMap *mapping, char *DevUDN, char *ServiceID);
 int CancelMappingExpiration(int eventId);
 void DeleteAllPortMappings(void);
-int AddNewPortMapping(struct Upnp_Action_Request *ca_event, char* new_enabled, int leaseDuration,
+int AddNewPortMapping(struct Upnp_Action_Request *ca_event, char* new_enabled, long int leaseDuration,
                      char* new_remote_host, char* new_external_port, char* new_internal_port,
                      char* new_protocol, char* new_internal_client, char* new_port_mapping_description,
                      int is_update);
